+++ R/generics.R
@@ -384,8 +384,10 @@ autocast_ab_to_tensors <- function()
 #' @param shape an integer vector, tensor, or `tf.TensorShape`. Can contain up
 #'   to 1 unspecified dimension, encoded as a `-1` or `NA`. This will reshape
 #'   `x` using row-major (C-style) semantics. It will prefer reshaping using
-#'   non-graph operations if possible, but will dispatch to `tf$reshape()`
-#'   otherwise.
+#'   non-graph operations if possible, but will otherwise invoke `tf$reshape()`.
+#'   If `x` is a scalar and the requested `shape` is fully defined or a tensor,
+#'   the value of `x` will be recycled to fill a tensor of the requested shape
+#'   (it will dispatch to `tf$fill()`).
 #' @param name `NULL` or a string. Useful for debugging in graph mode, ignored
 #'   while in eager mode.
 #'
@@ -408,12 +410,17 @@ as_tensor.default <- function(x, dtype = NULL, ..., shape = NULL, name = NULL) {
     shape <- as.integer.tensorflow.python.framework.tensor_shape.TensorShape(
       tensorflow::shape(dims = shape))
 
-    if (any(unspecified <- is.na(shape)))
-      shape[unspecified] <- -1L
+    if (partially_defined_shape <- any(undefined <- is.na(shape)))
+      shape[undefined] <- -1L
 
-    if(identical(dim(x), shape))
+    # if x is scalar and (shape fully defined or a tensor): tf$fill()
+    # else if x is R atomic or np.ndarray: np$reshape()
+    # else: tf$reshape()
+
+    if(identical(dx <- dim(x), shape))
       shape <- NULL
-    else if (is.array(x) || is.atomic(x) || inherits(x, "numpy.ndarray")) {
+    else if ((is.atomic(x) || inherits(x, "numpy.ndarray")) &&
+             !is_scalar(x)) {
       # prefer reshaping off the graph if possible
       np <- import("numpy", convert = FALSE)
       x <- np$reshape(x, shape, "C")
@@ -424,8 +431,13 @@ as_tensor.default <- function(x, dtype = NULL, ..., shape = NULL, name = NULL) {
   x <- tf$convert_to_tensor(x, dtype_hint = dtype, name = name)
   if (!is.null(dtype))
     x <- tf$cast(x, dtype, name = name)
-  if(!is.null(shape))
-    x <- tf$reshape(x, shape, name = name)
+
+  if (!is.null(shape)) {
+    if (is_scalar(x) && (is_tensor(shape) || !partially_defined_shape))
+      x <- tf$fill(shape, tf$squeeze(x), name = name)
+    else
+      x <- tf$reshape(x, shape, name = name)
+  }
 
   x
 }